successfully added and tested the post and get request through postman
